End of section five

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -9,6 +9,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -46,16 +47,17 @@ namespace AuctionService.Controllers
             if (auction == null) return NotFound();
             return _mapper.Map<AuctionDto>(auction);
         }
+
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAutionDto createAutionDto)
         {
             var auction = _mapper.Map<Auction>(createAutionDto);
 
-            //Todo: add current user as a seller 
-            auction.Seller = "testuser";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
-            
+
             //MassTransit kullanarak “AuctionCreated” adlı bir event gönderiyoruz
             //Yeni bir auction oluşturuldu!” bilgisini sistemin diğer mikroservislerine yani SearchService e gönderiyoruz.
             var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -64,16 +66,19 @@ namespace AuctionService.Controllers
             var result = await _context.SaveChangesAsync() > 0;
 
             if (!result) return BadRequest("Failed to save changes to auction database");
-            return CreatedAtAction(nameof(GetAuctionById), new { id = auction.Id },newAuction);
+            return CreatedAtAction(nameof(GetAuctionById), new { id = auction.Id }, newAuction);
 
         }
+
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
             var auction = await _context.Auctions.Include(a => a.Item).FirstOrDefaultAsync(x => x.Id == id);
             if (auction == null) return NotFound();
 
-            //Todo: check seller == username
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -89,13 +94,15 @@ namespace AuctionService.Controllers
             return BadRequest("Failed saving changes to auction database");
 
         }
+        
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
             if (auction == null) return NotFound();
 
-            //todo: check seller == username
+            if(auction.Seller != User.Identity.Name) return Forbid();
             _context.Auctions.Remove(auction);
 
             //MassTransit kullanarak “AuctionDeleted” adlı bir event gönderiyoruz
